refactor: Move network exceptions to their own submodule
